Allows compiling with ambiguous type conversion warnings enabled.

+++ DMTabBar/DMTabBar/DMTabBar.m
@@ -133,7 +133,7 @@
         totalWidth += item.itemWidth > 0 ? item.itemWidth : kDMTabBarItemWidth;
     }];
     
-    __block CGFloat offset_x = floorf((NSWidth(self.bounds)-totalWidth)/2.0f);
+    __block CGFloat offset_x = floor((NSWidth(self.bounds) - totalWidth) / 2.0f);
     [self.tabBarItems enumerateObjectsUsingBlock:^(DMTabBarItem* tabBarItem, NSUInteger idx, BOOL *stop) {
         CGFloat itemWidth = tabBarItem.itemWidth > 0 ? tabBarItem.itemWidth : kDMTabBarItemWidth;
         tabBarItem.tabBarItemButton.frame = NSMakeRect(offset_x, NSMinY(self.bounds),
